+++ src/metrics.jl
@@ -194,7 +194,7 @@ function pdist!(r, sa2, sb2)
     r
 end
 function pairwise!(r::AbstractMatrix, dist::SqEuclidean, a::AbstractMatrix)
-    m::Int, n::Int = get_pairwise_dims(r, a)
+    m, n = get_pairwise_dims(r, a)
     At_mul_B!(r, a, a)
     sa2 = sumsq_percol(a)
     for j = 1 : n
@@ -211,7 +211,7 @@ end
 
 # Euclidean
 function pairwise!(r::AbstractMatrix, dist::Euclidean, a::AbstractMatrix, b::AbstractMatrix)
-    m::Int, na::Int, nb::Int = get_pairwise_dims(r, a, b)
+    m, na, nb = get_pairwise_dims(r, a, b)
     At_mul_B!(r, a, b)
     sa2 = sumsq_percol(a)
     sb2 = sumsq_percol(b)
@@ -224,7 +224,7 @@ function pairwise!(r::AbstractMatrix, dist::Euclidean, a::AbstractMatrix, b::Abs
     r
 end
 function pairwise!(r::AbstractMatrix, dist::Euclidean, a::AbstractMatrix)
-    m::Int, n::Int = get_pairwise_dims(r, a)
+    m, n = get_pairwise_dims(r, a)
     At_mul_B!(r, a, a)
     sa2 = sumsq_percol(a)
     for j = 1 : n
@@ -242,7 +242,7 @@ end
 
 # CosineDist
 function pairwise!(r::AbstractMatrix, dist::CosineDist, a::AbstractMatrix, b::AbstractMatrix)
-    m::Int, na::Int, nb::Int = get_pairwise_dims(r, a, b)
+    m, na, nb = get_pairwise_dims(r, a, b)
     At_mul_B!(r, a, b)
     ra = sqrt!(sumsq_percol(a))
     rb = sqrt!(sumsq_percol(b))
@@ -284,8 +284,3 @@ end
 function pairwise!(r::AbstractMatrix, dist::CorrDist, a::AbstractMatrix)
     pairwise!(r, CosineDist(), _centralize_colwise(a))
 end
-
-
-
-
-
